Rename Python modules according to review
