+++ lib/spack/spack/operating_systems/_operating_system.py
