+++ lib/spack/spack/operating_systems/mac_os.py
@@ -8,7 +8,7 @@ import platform as py_platform
 from spack.util.executable import Executable
 from spack.version import Version
 
-from ._base_classes import OperatingSystem
+from ._operating_system import OperatingSystem
 
 
 # FIXME: store versions inside OperatingSystem as a Version instead of string
