+++ lib/spack/spack/platforms/_platform.py
