+++ lib/spack/spack/platforms/linux.py
@@ -9,7 +9,7 @@ import archspec.cpu
 import spack.target
 from spack.operating_systems.linux_distro import LinuxDistro
 
-from ._base_classes import Platform
+from ._platform import Platform
 
 
 class Linux(Platform):
